+++ sources/loaders/raw_loader/raw_writer.c
@@ -56,7 +56,7 @@ int RAW_libWrite_DiskFile(HXCFLOPPYEMULATOR* floppycontext,FLOPPY * floppy,char 
 		{
 			for(i=0;i<(int)floppy->floppyNumberOfSide;i++)
 			{
-				sprintf(tmp_str,"track:%.2d:%d file offset:0x%.6x, sectors: ",j,i,ftell(rawfile));
+				sprintf(tmp_str,"track:%.2d:%d file offset:0x%.6x, sectors: ",j,i,(unsigned int)ftell(rawfile));
 
 				log_str=0;
 				log_str=realloc(log_str,strlen(tmp_str)+1);
